💥 Replace existing implementations
