+++ .gitignore
@@ -1,4 +1,4 @@
 .idea
-/.php_cs.cache
-/vendor/
+.php_cs.cache
 composer.lock
+vendor/
