+++ .php_cs.dist
@@ -1,0 +1,20 @@
+<?php
+/*
+ * This document has been generated with
+ * https://mlocati.github.io/php-cs-fixer-configurator/#version:2.15.3|configurator
+ * you can change this configuration by importing this file.
+ */
+return PhpCsFixer\Config::create()
+    ->setRiskyAllowed(true)
+    ->setRules([
+        '@PSR2' => true,
+        '@PhpCsFixer' => true,
+        '@PhpCsFixer:risky' => true,
+        'declare_strict_types' => true,
+        'native_function_invocation' => true,
+        'yoda_style' => ['always_move_variable' => true, 'equal' => false, 'identical' => false],
+    ])
+    ->setFinder(PhpCsFixer\Finder::create()
+        ->exclude(['spec', 'vendor'])
+        ->in(__DIR__)
+    );
