--- phpmd.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<ruleset>
-    <rule ref="rulesets/cleancode.xml"/>
-    <rule ref="rulesets/codesize.xml"/>
-    <rule ref="rulesets/controversial.xml"/>
-    <rule ref="rulesets/design.xml"/>
-    <rule ref="rulesets/unusedcode.xml"/>
-    <rule ref="rulesets/naming.xml">
-        <exclude name="ShortMethodName"/>
-        <exclude name="ShortVariable"/>
-    </rule>
-    <rule ref="rulesets/naming.xml/ShortMethodName">
-        <properties>
-            <property name="exceptions" value="id"/>
-        </properties>
-    </rule>
-    <rule ref="rulesets/naming.xml/ShortVariable">
-        <properties>
-            <property name="exceptions" value="id"/>
-        </properties>
-    </rule>
-</ruleset>
