+++ src/Assert/AssertionFailed.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WebGarden\Model\Assert;
+
+use Assert\InvalidArgumentException;
+
+class AssertionFailed extends InvalidArgumentException
+{
+}
