--- src/Model/Assert/Assertion.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace WebGarden\Model\Assert;
-
-class Assertion extends \Assert\Assertion
-{
-    protected static $exceptionClass = AssertionException::class;
-}
