--- src/Model/Assert/AssertionException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace WebGarden\Model\Assert;
-
-use Assert\InvalidArgumentException;
-
-/**
- * @deprecated Will be renamed in the next major version
- */
-class AssertionException extends InvalidArgumentException
-{
-}
