--- src/Model/Assert/LazyAssertionException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace WebGarden\Model\Assert;
-
-/**
- * @deprecated Will be renamed in the next major version
- */
-class LazyAssertionException extends \Assert\LazyAssertionException
-{
-}
