--- src/Model/Entity/Identifiable.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace WebGarden\Model\Entity;
-
-/**
- * @deprecated Will be removed in the next major version
- */
-interface Identifiable
-{
-    /**
-     * Return identity of the entity.
-     *
-     * @return \WebGarden\Model\ValueObject\ValueObject
-     */
-    public function id();
-}
