+++ src/ValueObject/Number/Integer.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WebGarden\Model\ValueObject\Number;
+
+use WebGarden\Model\ValueObject\SingleValue;
+
+class Integer extends SingleValue
+{
+    protected function assertThat($value)
+    {
+        return parent::assertThat($value)->integer();
+    }
+}
